fixup! JAMES-2344 Strong type QuotaSize and QuotaCount to allow representing more cases

+++ mailbox/api/src/test/java/org/apache/james/mailbox/quota/QuotaCountTest.java
@@ -24,7 +24,7 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 
 public class QuotaCountTest implements QuotaValueTest<QuotaCount> {
 
-    public QuotaCount instance(int value) {
+    public QuotaCount instance(long value) {
         return QuotaCount.count(value);
     }
